Move common.generators to common.data.generators
