+++ src/rastervision/common/data/generators.py
